feat: 리뷰 내용 반영 (AtomicInteger 제거)

+++ src/main/kotlin/step3/domain/car/CarFactory.kt
@@ -1,14 +1,12 @@
 package step3.domain.car
 
-import java.util.concurrent.atomic.AtomicInteger
-
 object CarFactory {
 
     private const val DEFAULT_POSITION = 0
 
     fun createCar(name: String, position: Int = DEFAULT_POSITION): Car = Car(
         name = name,
-        position = AtomicInteger(position),
+        position = position,
     )
 
     fun createCars(
